Add logging for any endpoint of any controller
