+++ src/main/java/co/raccoons/bookkeeper/BookkeeperOperationStatus.java
@@ -3,11 +3,13 @@ package co.raccoons.bookkeeper;
 import javax.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.ToString;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 @ResponseBody
 @AllArgsConstructor
 @Getter
+@ToString
 public final class BookkeeperOperationStatus {
 
     @NotNull
